moved input handler into its own file
